+++ mayan/apps/document_states/views/__init__.py
@@ -1,0 +1,3 @@
+from .workflow_instance_views import *  # NOQA
+from .workflow_proxy_views import *  # NOQA
+from .workflow_views import *  # NOQA