+++ HISTORY.rst
@@ -8,6 +8,7 @@
   during a check.
 * Switch to full app paths.
 * Split document app models into separate modules.
+* Split workflow views into separate modules.
 
 3.1.11 (2019-04-XX)
 ===================